cosmetic changes
adding and reducing padding to make the whole thing a bit neater.

Had to hardcode a few things in v5 ot make it easier to use some of the new tag styles in v6.

Ceased page was showing the added test banner so this only appears when adjusting the cease/defer and not from the nhs search box