Stop watching html files
Currently files with the extension `.html` are being watched, which means that whenever they are edited, the nodemon server is restarted.

When combined with the in-memory session storage (which is currently the default), this means that when editing any html file, all data previously entered by the user is wiped. This makes local development of prototypes where data is stored quite tricky.

Since nunjucks is now configured to not cache the files (see #286), we no longer need to watch the HTML files and reload the server, as any changes to the content will still be shown.

Fixes #287