Set Strict-Transport-Security header in production
This will ensure that browsers automatically load the prototype over https, even if the `http://` prefix is typed into the URL bar.

It does require `NODE_ENV` to be set to `production`, however many platforms such as Heroku will [set this by default](https://devcenter.heroku.com/changelog-items/688) for Node.js apps.

Once this is set you can also submit the domain to https://hstspreload.org which will ensure the browsers at it to their list of domains to always use https on.