Add instructions if you run npm start
Currently if you run `npm start`, which is the common default start command for Node.js apps, then in the console you don’t see any messages and it looks like nothing is working.

However it is working, just on the uncommon default port of 2000 (set in `config.js`).

This adds a console message to say:

> Running at http://localhost:2000/
>
> If you are working on this prototype now, press Control-C to stop the server and
> type npm run watch instead to run the prototype in development mode.

This message is not shown when running `npm run watch`, as there the BrowserSync service will display the proxied localhost port to view the prototype on.