Remove example page
This removes the example page, and the links to it. The example page has a number of example content components on it, including headings, lede text, figures, tables, tabs, lists, warning callouts and page review date.

However we think it’d be more useful to point people at the NHS Design system to browse the styles and components, especially as the guidance on how and when to use them is there too.