Update the release process description (#445)
Updated notes on the release process.

Probably this shouldn’t be in the README for the kit itself (as that
might be confusing for anyone who gets a copy of it with their
prototype) - but there is an existing note saying it’s only for the
team. Perhaps in future it could get moved somewhere else.