Use strict https headers and redirect from http to https in production mode (#410)
This aims to avoid browser warnings if users accidentally visit a
prototype in non-secure mode.

It does this by adding the [`Strict-Transport-Security` HTTP
header](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Strict-Transport-Security)
and `Content-Security-Policy` to `upgrade-insecure-requests`.

This will ensure that all subsequent requests will automatically use
https, even if the `http://` prefix is typed into the URL bar.

Additionally, the kit will automatically respond with a redirect to the
https url whenever insecure requests are made.

This only happens when
[`NODE_ENV`](https://nodejs.org/en/learn/getting-started/nodejs-the-difference-between-development-and-production)
is set to `production`, to avoid any issues when running the prototype
over localhost

Once your prototype is running with the header set, you can also submit
the domain to https://hstspreload.org/ (although this doesn't work for
subdomains) which will ensure the browsers at it to their list of
domains to always use https on without even having to visit it first.

Fixes #141

## Checklist

- [x] CHANGELOG entry

---------

Co-authored-by: Colin Rotherham <work@colinr.com>