Remove stub test for app filters (#509)
This test simply asserts that there are no filters set up in
`app/filters.js`.

This means that if you add any custom filters, the test will fail.
Whilst this could be useful as a prompt for users to create tests for
their own filters, we don't want to require all users to do this, and
it's likely unnecessary for a prototype.