Open preview in a new browser tab instead of an preview (#628)
This defaults the prototype to running in a new browser tab, instead of
a tab within Codespaces itself.

The advantage to this is that if you refresh it using Control-R, it'll
only refresh the browser tab and not the whole of codespaces. It also
makes it a bit easier to test the responsiveness by resizing the window.

The downside is that depending on your browser settings you might get a
popup warning first:

<img width="1040" height="555" alt="Screenshot 2025-09-10 at 10 30 03"
src="https://github.com/user-attachments/assets/5b56971b-9995-4655-9297-1b5110484e66"
/>